discrete RL bugfix

+++ code/learning/ppo.py
@@ -73,7 +73,7 @@ class PPO(nn.Module):
 			done_lst.append([done_mask])
 			
 		s,a,r,s_prime,done_mask, prob_a = torch.tensor(s_lst, dtype=torch.float), torch.tensor(a_lst), \
-										  torch.tensor(r_lst), torch.tensor(s_prime_lst, dtype=torch.float), \
+										  torch.tensor(r_lst, dtype=torch.float), torch.tensor(s_prime_lst, dtype=torch.float), \
 										  torch.tensor(done_lst, dtype=torch.float), torch.tensor(prob_a_lst)
 		self.data = []
 		return s, a, r, s_prime, done_mask, prob_a
